[pydrake] Factor out the std::monostate support for reuse
