+++ bindings/pydrake/geometry_py_scene_graph.cc
@@ -5,30 +5,12 @@
 
 #include "drake/bindings/pydrake/common/default_scalars_pybind.h"
 #include "drake/bindings/pydrake/common/deprecation_pybind.h"
+#include "drake/bindings/pydrake/common/monostate_pybind.h"
 #include "drake/bindings/pydrake/common/type_pack.h"
 #include "drake/bindings/pydrake/common/value_pybind.h"
 #include "drake/bindings/pydrake/documentation_pybind.h"
 #include "drake/geometry/geometry_frame.h"
 #include "drake/geometry/scene_graph.h"
-
-// TODO(SeanCurtis-TRI) When pybind issue 3019 gets resolved, we won't need to
-//  define this locally anymore. In fact, it will probably cause link errors.
-namespace pybind11 {
-namespace detail {
-template <>
-struct type_caster<std::monostate> {
- public:
-  PYBIND11_TYPE_CASTER(std::monostate, _("None"));
-
-  bool load(handle src, bool) { return src.ptr() == Py_None; }
-
-  static handle cast(
-      std::monostate, return_value_policy /* policy */, handle /* parent */) {
-    Py_RETURN_NONE;
-  }
-};
-}  // namespace detail
-}  // namespace pybind11
 
 namespace drake {
 namespace pydrake {
